Resolve SPIKE-13.1 (it was a UI-only enable button); fix the six 3.10 mypy errors
ADR-0008 closed as Resolved: the owner clicked an 'Enable GitHub Actions
in this repository' control that neither settings review surfaced, and
the very next push (af6036b) triggered the repository's first-ever
Actions run within seconds, byte-identical workflow, no other change.
The API had reported enabled: true / allowed_actions: all the whole time
-- an API/backend disagreement worth a low-priority bug report to GitHub,
recorded in the ADR.

First run outcome: test (3.11) and test (3.12) fully green on clean
Ubuntu -- lint, format, citations, all 1082 tests. test (3.10) failed
only at mypy, on six numpy-stub-generation typing divergences (3.10
resolves an older numpy than 3.11/3.12 at install time, and dtype
inference differs between stub generations). Fixed by making all six
call sites explicitly float64-typed -- np.linspace dtype in
array/beamform.py (x2) and viz/slices.py (x2), an NDArray[np.float64]
annotation in propagate/retarded.py, and a concrete asarray on
viz/patterns.py's np.maximum return -- the fix the FP64-everywhere rule
asks for anyway, rather than weakening or version-pinning the CI gate.

All five local gates green. This push is the T-13.2 green-run attempt.

+++ src/gwtb/array/beamform.py
@@ -184,7 +184,7 @@ def beamwidth_3db(
     ax = np.asarray(axis, dtype=np.float64)
     ax = ax / np.linalg.norm(ax)
 
-    theta = np.linspace(-np.pi / 2.0, np.pi / 2.0, 200001)
+    theta = np.linspace(-np.pi / 2.0, np.pi / 2.0, 200001, dtype=np.float64)
     pattern = _pattern_along_broadside_sweep(pos, w, wavelength, ax, theta)
     peak = np.max(pattern)
     half_power = peak / np.sqrt(2.0)
@@ -244,7 +244,7 @@ def peak_sidelobe_level(
     ax = np.asarray(axis, dtype=np.float64)
     ax = ax / np.linalg.norm(ax)
 
-    theta = np.linspace(-np.pi / 2.0, np.pi / 2.0, 400001)
+    theta = np.linspace(-np.pi / 2.0, np.pi / 2.0, 400001, dtype=np.float64)
     pattern = _pattern_along_broadside_sweep(pos, w, wavelength, ax, theta)
     peak = np.max(pattern)
     peak_idx = int(np.argmax(pattern))
